[Android] Move BraveUrlConstants out of embedder_support component
`BraveUrlConstants` does not extend anything and thus does not belong to `embedder_support` component.
