+++ android/brave_java_sources.gni
@@ -540,6 +540,7 @@ brave_java_base_sources = [
   "//brave/android/java/org/chromium/base/BraveFeatureList.java",
   "//brave/android/java/org/chromium/base/BravePreferenceKeys.java",
   "//brave/android/java/org/chromium/base/BraveReflectionUtil.java",
+  "//brave/android/java/org/chromium/base/BraveUrlConstants.java",
   "//brave/android/java/org/chromium/base/Callbacks.java",
 ]
