+++ android/java/org/chromium/base/BraveUrlConstants.java
@@ -3,11 +3,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-package org.chromium.components.embedder_support.util;
+package org.chromium.base;
 
-/**
- * Extension of UrlConstants.java
- */
+/** Extension of UrlConstants.java */
 public class BraveUrlConstants {
     public static final String BRAVE_SCHEME = "brave";
 }
