+++ components/embedder_support/android/java_sources.gni
@@ -3,7 +3,4 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at https://mozilla.org/MPL/2.0/.
 
-brave_components_embedder_support_java_sources = [
-  "//brave/components/embedder_support/android/java/src/org/chromium/components/embedder_support/util/BraveUrlConstants.java",
-  "//brave/components/embedder_support/android/java/src/org/chromium/components/embedder_support/view/BraveContentView.java",
-]
+brave_components_embedder_support_java_sources = [ "//brave/components/embedder_support/android/java/src/org/chromium/components/embedder_support/view/BraveContentView.java" ]
